Allow for interrupting AMQPChannel._wait_on_frame
Required to address issue #38. Adds a ``threading.Event`` that is checked in the ``AMQPChannel._wait_on_frame`` method which can be used to interrupt a blocking wait. Also implemented a check to keep the debug logging in the code while optimizing to prevent the performance overhead normally seen when debug logging without the logger actively set to debug. This value is cached, so I should potentially create a way to clear the internal cached representation for apps that change logging config on the fly.